# backend/docs/swagger.yaml
components:
  schemas:
    api.HealthResponse:
      properties:
        status:
          example: healthy
          type: string
      type: object
    data:
      properties:
        data:
          $ref: '#/components/schemas/api.HealthResponse'
      type: object
...
    handler.SavedQueryRequest:
      properties:
        connection_id:
          example: 9f1c7d2e4a6b4e89b88ad5f356bf7312
          type: string
        sql_text:
          example: SELECT * FROM users WHERE status = 'active'
          type: string
        tags:
          example: users,report
          type: string
        title:
          example: Active users
          type: string
      required:
      - sql_text
      - title
      type: object
...
    response.APIError:
      properties:
        code:
          type: string
        message:
          type: string
        position:
          type: integer
      type: object
...
    response.ErrorEnvelope:
      properties:
        data:
          description: always null
        error:
          $ref: '#/components/schemas/response.APIError'
        meta:
          additionalProperties: {}
          type: object
        success:
          example: false
          type: boolean
      type: object
...
info:
  description: Local-first database GUI backend. All responses use the standard envelope
    (success, data, error, meta).
  title: DataDeck API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /connections/{id}/schemas:
    get:
      description: Returns a database-neutral schema tree (schemas, tables, columns,
        primary keys, foreign keys, indexes) for a stored connection.
      parameters:
      - description: Connection id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Not Found
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Bad Gateway
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Gateway Timeout
      summary: Get database schema
      tags:
      - schema
...
  /queries/saved:
    get:
      description: Returns saved queries newest-updated first, optionally filtered
        by connection. (M3 API extension.)
      parameters:
      - description: Filter by connection id
        in: query
        name: connection_id
        schema:
          type: string
      - description: Page number (1-based)
        in: query
        name: page
        schema:
          type: integer
      - description: Page size (max 200)
        in: query
        name: page_size
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Internal Server Error
      summary: List saved queries
      tags:
      - saved-queries
    post:
      description: Persists a SQL snippet. The SQL is stored, never executed. (PRD
        endpoint.)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.SavedQueryRequest'
              description: Saved query
              summary: request
        description: Saved query
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Not Found
      summary: Create a saved query
      tags:
      - saved-queries
  /queries/saved/{id}:
    delete:
      parameters:
      - description: Saved query id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Not Found
      summary: Delete a saved query
      tags:
      - saved-queries
    get:
      parameters:
      - description: Saved query id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Not Found
      summary: Get a saved query
      tags:
      - saved-queries
    put:
      description: Replaces the mutable fields of a saved query. (M3 API extension.)
      parameters:
      - description: Saved query id
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.SavedQueryRequest'
              description: Saved query
              summary: request
        description: Saved query
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Not Found
      summary: Update a saved query
      tags:
      - saved-queries
...
servers:
- url: /api/v1